requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
cycler==0.10.0
dateparser==1.0.0
debugpy==1.3.0
decorator==5.0.9
idna==3.2
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
multidict==5.1.0
numpy==1.21.1
pandas==1.3.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
plotly==5.1.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
pyserial==3.5
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.1.0
regex==2021.7.6
requests==2.26.0
scipy==1.7.0
six==1.16.0
tclab==0.4.9
tenacity==8.0.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
wcwidth==0.2.5
websockets==9.1
yarl==1.6.3
